% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGSE_v1.95_new.R
\name{error_minimize_raw}
\alias{error_minimize_raw}
\title{Minimize the Error for Raw K-mer Frequency Fitting}
\usage{
error_minimize_raw(
  tooptimize,
  x,
  end,
  xfit,
  xfit_left,
  xfit_right,
  d,
  min_valid_pos,
  itr
)
}
\arguments{
\item{tooptimize}{A numeric vector containing the scale factors to optimize.}

\item{x}{A numeric vector representing the histogram replicated from \code{d}.}

\item{end}{An integer indicating the right-side position for fitting.}

\item{xfit}{x-values to get the skew normal distribution}

\item{xfit_left}{A numeric value for the left-side position to calculate initial mean and standard deviation.}

\item{xfit_right}{A numeric value for the right-side position to calculate initial mean and standard deviation.}

\item{d}{A data frame representing the observed k-mer frequencies that will be fitted.}

\item{min_valid_pos}{An integer indicating the left-side position from which the observed k-mer frequencies will be fitted.}

\item{itr}{An integer representing the iteration count.}
}
\value{
A numeric value representing the minimized error.
}
\description{
This function minimizes the error for raw k-mer frequency fitting by adjusting the mean, standard deviation, and scaling factors.
}
