% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initial_count_recover}
\alias{initial_count_recover}
\title{Recover Initial K-mer Count}
\usage{
initial_count_recover(d0)
}
\arguments{
\item{d0}{A data frame representing the initial k-mer count from software like Jellyfish.}
}
\value{
A data frame with recovered k-mer counts.
}
\description{
This function recovers the initial k-mer count, making the k-mer frequency consecutive.
}
\examples{
d0 <- data.frame(V1 = c(1, 2, 4), V2 = c(100, 200, 300))
dr <- initial_count_recover(d0)
}
