% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_start_time}
\alias{initialize_start_time}
\title{Initialize Start Time}
\usage{
initialize_start_time()
}
\value{
The start time of the process.
}
\description{
This function initializes the start time for a process.
}
