% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGSE_v1.95_new.R
\name{kmer_count_modify_raw}
\alias{kmer_count_modify_raw}
\title{Modify Raw K-mer Frequency Before Fitting}
\usage{
kmer_count_modify_raw(start, end, left_right, histx)
}
\arguments{
\item{start}{An integer indicating the starting position (does not include the peak position).}

\item{end}{An integer indicating the ending position (does not include the peak position).}

\item{left_right}{An integer indicating whether to modify the left part (0) or the right part (1).}

\item{histx}{A data frame representing the raw k-mer histogram.}
}
\value{
A modified data frame with adjusted raw k-mer frequencies.
}
\description{
This function modifies the raw k-mer frequency before fitting, adjusting either the left or right part based on the peak position.
}
